refactor(svelte2tsx): 式・命名・ソース走査ヘルパを utils/ へ切り出す
式レンジ/テキストと TS アサーション剥がし (utils/expr.rs)、生成変数名
(utils/names.rs)、タグ位置のソース走査 (utils/source.rs) を純粋移動。
対応するユニットテストも移設。出力・ロジックの変更なし。
