+++ crates/rsvelte_core/src/svelte2tsx/template/utils/mod.rs
@@ -1,0 +1,5 @@
+//! Low-level helpers shared by the template handlers.
+
+pub(super) mod expr;
+pub(super) mod names;
+pub(super) mod source;
